backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
python-multipart>=0.0.12,<1
openai>=1.57.0,<2
google-generativeai>=0.8.0,<1
pydantic-settings>=2.6.0,<3
python-dotenv>=1.0.0,<2
pypdf>=5.0.0,<6
numpy>=1.26.0,<3
aiosqlite>=0.20.0,<1
python-docx>=1.1.0,<2
sse-starlette>=2.0.0,<3
